src/requirements.txt
numpy == 2.2.6
pandas == 2.3.3
scikit-learn == 1.8.0
Flask == 3.1.2
psycopg2-binary == 2.9.11
bcrypt == 5.0.0
groq

